Async replay working (a little too well)
It allows the UART to "disconnect" if you cross the point
when it was connected. If you then advance forward, the
UART peripheral is still in recv(). This is fine so long
as you seek forwards again; however if you instead execute
forward, the UART will never 'reconnect' (not that I'm sure
about best behavior there either, since the other end would
need to somehow divine the need to reconnect).

This shall require a little more thought