Add speed throttling: simulated core freq control
The posix implementation is added blind / untested.
The apple implementation is tested and seems to work.

No optimizations, etc have been made for performance really. Most
critically state tracking is still always on which is going to be
inefficient / at odds with trying to get any kind of high-speed or timing
critical application working. There is also the issue of lock contention
in the state tracking code, which further hurts performance.

With all that said, my otherwise unloaded mac pro can consistantly run at
about 700 Hz on a 2.7 GHz i7 core. Only order 4 million or so overhead--
lots of room for improvement! :)