Print loose approximation of simulator speed
It's nice to have a mental ballpark figure of simulator performance, so
we now print the "average simulated frequency" on exit. This is a really
rough number, it's simply (cycle count / execution time).

The syscall / context switch overhead of the current speed limiting code
causes the simulator to be consistently slower than the requested speed.
While this isn't really surprising to see, it's nice to see that problem
validated with real numbers. I'll have to look into how other emulators
achieve tighter timing at some point in the future --- I'm guessing they
either do something adaptive, or a simpler would be to track sum of
execution time, shoot for an average, and consistently undershoot the
sleep call, that would more or less work things out.