Allow non-main thread to call terminate during startup
During startup, the array of threads may have invliad pthread_id
members. Add an 'active' element to indicate that a thread has been
started. This flag is used to avert an invalid join if a thread hasn't
started yet (occurs when the simulator is terminated by an error in a
thread that it is spawning).