re-license as dual MIT/Apache2
This re-licensing is done with the written permission of all contributors.

The express intent is to re-license the entirity of this work as a dual
licensed under both MIT and/or Apache2. Any headers or notices that may
have accidentally been missed are incidental and should not be taken as
an attempt to preserve prior licensing terms.